Fix problem with restart loop
In case action exits normally, the loop just exits
without restarting. And this happens when amqp encounters any abnormal
exception, because it first closes the connection, which triggers a
write to the TMVar, on which action is waiting

Solution: Instead of the TMVar just being there for decoration, actually
use it to catch the asynchronous exception thrown from the channel
closed exception handler and rethrow it on the action thread, such that
in *all* cases, action won't exit normally. Why not connection closed handler?
Because there's no exception available there. We'll just assume the connection
always gets closed when the channel does (which upon second thought
doesn't sound so good).